Refactor menu Work Item and Fix Corrections
1. remove this method, it is not return an object. (Fixed)
2. please rename this method clickWorkItemMenu (Fixed)
3. Refactor Menu Work Item
