+++ src/main/java/org/umssdiplo/automationv01/core/managepage/menuheader/workItemsMenu/WorkItemsMenu.java
@@ -1,8 +1,9 @@
-package org.umssdiplo.automationv01.core.managepage.workItem;
+package org.umssdiplo.automationv01.core.managepage.menuheader.workItemsMenu;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.umssdiplo.automationv01.core.managepage.BasePage;
+import org.umssdiplo.automationv01.core.managepage.workItem.WorkItemList;
 import org.umssdiplo.automationv01.core.utils.CommonEvents;
 
 public class WorkItemsMenu extends BasePage {
